pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clustered-ep-risk-mapper"
version = "0.1.0"
description = "Interactive thermal and EMI risk mapper for clustered electric propulsion layouts."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "plotly>=5.24",
  "pyyaml>=6.0",
  "streamlit>=1.39",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = [
  "src",
]
testpaths = [
  "tests",
]
